@@ -1,23 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import 'package:todo_app/screens/add_task/add_task_screen.dart';
-import 'package:todo_app/screens/tasks/tasks_screen.dart';
 
-import 'models/task.dart';
+import 'src/app.dart';
 
 void main() {
-  runApp(MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return ChangeNotifierProvider(
-      create: (context) => TaskModel(),
-      child: MaterialApp(
-        title: 'To Do List',
-        home: TasksScreen(),
-      ),
-    );
-  }
+  runApp(const App());
 }
