@@ -1,33 +1,33 @@
-import 'package:flutter/material.dart';
-import 'package:todo_app/screens/add_task/components/body.dart';
+// import 'package:flutter/material.dart';
+// import 'package:todo_app/screens/add_task/components/body.dart';
 
-class AddTaskScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.white,
-      appBar: AppBar(
-        elevation: 0,
-        centerTitle: true,
-        title: Text(
-          'New task',
-          style: TextStyle(color: Colors.black),
-        ),
-        backgroundColor: Colors.white,
-        actions: [
-          IconButton(
-            icon: Icon(
-              Icons.close,
-              color: Colors.black,
-            ),
-            onPressed: () {
-              Navigator.pop(context);
-            },
-            tooltip: 'Close',
-          ),
-        ],
-      ),
-      body: Body(),
-    );
-  }
-}
+// class AddTaskScreen extends StatelessWidget {
+//   @override
+//   Widget build(BuildContext context) {
+//     return Scaffold(
+//       backgroundColor: Colors.white,
+//       appBar: AppBar(
+//         elevation: 0,
+//         centerTitle: true,
+//         title: Text(
+//           'New task',
+//           style: TextStyle(color: Colors.black),
+//         ),
+//         backgroundColor: Colors.white,
+//         actions: [
+//           IconButton(
+//             icon: Icon(
+//               Icons.close,
+//               color: Colors.black,
+//             ),
+//             onPressed: () {
+//               Navigator.pop(context);
+//             },
+//             tooltip: 'Close',
+//           ),
+//         ],
+//       ),
+//       body: Body(),
+//     );
+//   }
+// }
