@@ -1,37 +1,37 @@
-import 'package:flutter/material.dart';
-import 'package:todo_app/screens/add_task/add_task_screen.dart';
-import 'package:todo_app/screens/tasks/components/body.dart';
+// import 'package:flutter/material.dart';
+// import 'package:todo_app/screens/add_task/add_task_screen.dart';
+// import 'package:todo_app/screens/tasks/components/body.dart';
 
-class TasksScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.lightBlue,
-      appBar: AppBar(
-        elevation: 0,
-        actions: [
-          Icon(
-            Icons.list,
-          ),
-        ],
-      ),
-      body: Body(),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          Navigator.push(
-            context,
-            MaterialPageRoute(
-              builder: (context) => AddTaskScreen(),
-            ),
-          );
-        },
-        child: Icon(
-          Icons.add,
-          color: Colors.white,
-          size: 32,
-        ),
-        backgroundColor: Colors.lightBlue,
-      ),
-    );
-  }
-}
+// class TasksScreen extends StatelessWidget {
+//   @override
+//   Widget build(BuildContext context) {
+//     return Scaffold(
+//       backgroundColor: Colors.lightBlue,
+//       appBar: AppBar(
+//         elevation: 0,
+//         actions: [
+//           Icon(
+//             Icons.list,
+//           ),
+//         ],
+//       ),
+//       body: Body(),
+//       floatingActionButton: FloatingActionButton(
+//         onPressed: () {
+//           Navigator.push(
+//             context,
+//             MaterialPageRoute(
+//               builder: (context) => AddTaskScreen(),
+//             ),
+//           );
+//         },
+//         child: Icon(
+//           Icons.add,
+//           color: Colors.white,
+//           size: 32,
+//         ),
+//         backgroundColor: Colors.lightBlue,
+//       ),
+//     );
+//   }
+// }
